# openapi.yaml
openapi: 3.0.0
info:
  title: University Leave Management System API
  description: API documentation for the University Leave Management System.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local development server
paths:
...
  /leave-requests:
    post:
      summary: Create Leave Request
      description: Submit a new leave request. Supports attachments (multipart/form-data).
      security:
        - BearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - leaveTypeId
                - startDate
                - endDate
                - totalDays
              properties:
                leaveTypeId:
                  type: integer
                  example: 1
                startDate:
                  type: string
                  format: date
                  example: '2026-06-15'
                endDate:
                  type: string
                  format: date
                  example: '2026-06-16'
                totalDays:
                  type: number
                  format: float
                  example: 2.0
                timeSlot:
                  type: string
                  enum: [full, morning, afternoon]
                  default: full
                reason:
                  type: string
                  example: Going on a family trip
                contactAddress:
                  type: string
                  example: 123 Main St
                contactPhone:
                  type: string
                  example: '0812345678'
                attachments:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Leave request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRequest'
        '400':
          description: Bad request (e.g. insufficient leave balance)
        '401':
          description: Unauthorized

    get:
      summary: Get My Leave Requests
      description: Returns leave requests submitted by the logged-in user.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of user's leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaveRequest'
        '401':
          description: Unauthorized
...
  /leave-types:
    get:
      summary: Get All Leave Types
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of leave types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaveType'

    post:
      summary: Create Leave Type (Admin Only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - code
              properties:
                name:
                  type: string
                  example: Sick Leave
                code:
                  type: string
                  enum: [sick, personal, vacation, maternity, paternity, childcare, ordination, military]
                  example: sick
                description:
                  type: string
                defaultDays:
                  type: integer
                  example: 30
                requiresMedicalCert:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Created successfully
...
  /users:
    get:
      summary: Get All Users (Admin Only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      summary: Create User (Admin Only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - employeeId
                - email
                - password
              properties:
                employeeId:
                  type: string
                  example: EMP001
                email:
                  type: string
                  format: email
                  example: staff@university.edu
                password:
                  type: string
                  example: password123
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
                role:
                  type: string
                  enum: [employee, head, admin]
                  default: employee
                position:
                  type: string
                departmentId:
                  type: integer
                supervisorId:
                  type: integer
      responses:
        '201':
          description: User created successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        employeeId:
          type: string
          example: EMP001
        email:
          type: string
          format: email
          example: user@example.com
        firstName:
          type: string
          example: Somchai
        lastName:
          type: string
          example: Jaidee
        role:
          type: string
          enum: [employee, head, admin]
          example: employee
        position:
          type: string
          example: Lecturer
        departmentId:
          type: integer
          example: 2
        supervisorId:
          type: integer
          nullable: true
          example: 3
        phone:
          type: string
          nullable: true
          example: '0812345678'
        profileImage:
          type: string
          nullable: true
          example: uploads/profiles/profile-1-123456789.png
        signatureImage:
          type: string
          nullable: true
          example: uploads/profiles/sig-1-123456789.png
        startDate:
          type: string
          format: date
          nullable: true
          example: '2020-01-01'
        governmentDivision:
          type: string
          nullable: true
        documentNumber:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        affiliation:
          type: string
          nullable: true
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LeaveRequest:
      type: object
      properties:
        id:
          type: integer
          example: 12
        userId:
          type: integer
          example: 1
        leaveTypeId:
          type: integer
          example: 1
        startDate:
          type: string
          format: date
          example: '2026-06-15'
        endDate:
          type: string
          format: date
          example: '2026-06-16'
        totalDays:
          type: number
          format: float
          example: 2.0
        timeSlot:
          type: string
          enum: [full, morning, afternoon]
          example: full
        reason:
          type: string
          example: Sick with flu
        contactAddress:
          type: string
        contactPhone:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, confirmed, cancelled]
          example: pending
        approvedBy:
          type: integer
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        confirmedBy:
          type: integer
          nullable: true
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        confirmedNote:
          type: string
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        cancelReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LeaveType:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Sick Leave
        code:
          type: string
          example: sick
        description:
          type: string
        defaultDays:
          type: integer
          example: 30
        requiresMedicalCert:
          type: boolean
          example: false
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time